Move education to utils
